+++ include/reactphysics3d/containers/Stack.h
@@ -62,7 +62,8 @@ class Stack {
         void allocate(uint64 capacity) {
 
             // Make sure capacity is an integral multiple of alignment
-            capacity = std::ceil(capacity / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
+            capacity =
+                static_cast<uint64>(std::ceil(capacity / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT);
 
             T* newArray = static_cast<T*>(mAllocator.allocate(capacity * sizeof(T)));
             assert(newArray != nullptr);
