+++ src/memory/HeapAllocator.cpp
@@ -66,7 +66,8 @@ HeapAllocator::~HeapAllocator() {
 
         // Destroy the unit
         unit->~MemoryUnitHeader();
-        const size_t sizeHeader = std::ceil(sizeof(MemoryUnitHeader) / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
+        const size_t sizeHeader =
+            static_cast<size_t>(std::ceil(sizeof(MemoryUnitHeader) / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT);
         mBaseAllocator.release(static_cast<void*>(unit), unitSize + sizeHeader);
 
         unit = nextUnit;
@@ -321,7 +322,8 @@ void HeapAllocator::mergeUnits(MemoryUnitHeader* unit1, MemoryUnitHeader* unit2)
 // Reserve more memory for the allocator
 void HeapAllocator::reserve(size_t sizeToAllocate) {
 
-    const size_t sizeHeader = std::ceil(sizeof(MemoryUnitHeader) / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
+    const size_t sizeHeader =
+        static_cast<size_t>(std::ceil(sizeof(MemoryUnitHeader) / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT);
 
     // Allocate memory
     void* memory = mBaseAllocator.allocate(sizeToAllocate + sizeHeader);
